requirements.txt
email_validator
Flask
flask_login
flask_wtf
mysql-connector-python
werkzeug